+++ utils/logger.js
@@ -2,14 +2,14 @@
  * @Author: 渔火Arcadia  https://github.com/yhArcadia
  * @Date: 2026-08-09 19:36:23
  * @LastEditors: 渔火Arcadia
- * @LastEditTime: 2026-08-09 19:56:46
- * @FilePath: /Yunzai/plugins/who-are-you-plugin/utils/logger.js
+ * @LastEditTime: 2026-08-12 16:10:11
+ * @FilePath: /kh-plugin/utils/logger.js
  * @Description: 覆写logger，添加插件专属前缀。
  * 
  * Copyright (c) 2026 by 渔火Arcadia 1761869682@qq.com, All Rights Reserved. 
  */
 
-const PREFIX = '[who_are_you] ';
+const PREFIX = '[kh-plugin] ';
 
 function write(level, ...args) {
     const method = logger[level];